bugfix/deleting_job_session_renavigates

+++ app/src/main/java/com/example/nathan/loadtracker/ui/views/JobSessionsFragment.kt
@@ -4,11 +4,13 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.nathan.loadtracker.R
 import com.example.nathan.loadtracker.core.database.entities.JobSession
 import com.example.nathan.loadtracker.databinding.FragmentJobSessionsBinding
 import com.example.nathan.loadtracker.ui.arrayadapters.JobSessionAdapter
@@ -43,31 +45,53 @@ class JobSessionsFragment : Fragment() {
             arrayListOf(),
             { selectedJobSession ->
                 viewModel.selectJobSession(selectedJobSession.id)
-                findNavController().popBackStack()
+                if (!findNavController().popBackStack()) {
+                    findNavController().navigate(R.id.action_global_trackingSessionFragment)
+                }
             },
-            { jobSession ->
-                viewModel.deleteJobSession(jobSession)
-//                            AlertDialog
-//                                .Builder(requireContext())
-//                                .setMessage("Really?")
-//                                .setPositiveButton("Yep!") { dialog, _ ->
-//                                    launch {
-//                                        viewModel.deleteJobSession(jobSession)
-//                                        dialog.dismiss()
-//                                    }
-//                                }
-//                                .setNegativeButton("Nah") { dialog, _ ->
-//                                    dialog.dismiss()
-//                                }
-//                                .show()
+            { jobSession, isActiveSession ->
+                AlertDialog
+                    .Builder(requireContext())
+                    .setMessage("Really?")
+                    .setPositiveButton("Yep!") { dialog, _ ->
+                        if (isActiveSession) {
+                            dialog.dismiss()
+
+                            AlertDialog
+                                .Builder(requireContext())
+                                .setMessage(
+                                    "This session is currently active! If you close this now," +
+                                            " you will have to pick another session or create a new one before proceeding. " +
+                                            "Are you sure?"
+                                )
+                                .setPositiveButton("Yep!") { dialog, _ ->
+                                    viewModel.deleteJobSession(jobSession)
+                                    findNavController().clearBackStack(R.id.trackingSessionFragment)
+                                    dialog.dismiss()
+                                }
+                                .setNegativeButton("Nah") { dialog, _ -> dialog.dismiss() }
+                                .show()
+                        } else {
+                            viewModel.deleteJobSession(jobSession)
+                            dialog.dismiss()
+                        }
+                    }
+                    .setNegativeButton("Nah") { dialog, _ ->
+                        dialog.dismiss()
+                    }
+                    .show()
             }
         )
+
         binding.rvJobSessions.layoutManager = LinearLayoutManager(requireContext())
         binding.rvJobSessions.adapter = listAdapter
 
         viewLifecycleOwner.lifecycleScope.launch {
-            viewModel.allJobSessions.collect { jobs ->
-                listAdapter.jobSessions = jobs as ArrayList<JobSession>
+            viewModel.jobSessionModelFlow.collect { model ->
+                listAdapter.jobSessions =
+                    model.allJobSessions.map {
+                        Pair(it, it.id == model.activeJobSessionId)
+                    } as ArrayList<Pair<JobSession, Boolean>>
                 listAdapter.notifyDataSetChanged()
             }
         }
